Contest Vote Summary Report completed.  This required some modifications to core classes.  First, I made a connection between ContestCounty and its ResponseValues in the class using bags for hibernate.  I also added some methods to core classes (RDD-styled methods, no structural changes made to core classes except ContestCounty -> ResponseValue).
Note: I also noticed that the changes made to the vote entry form for persistence validation have broken it- it cannot persist.  My guess is this is ok because we are not done with this, but I figured I would note it because it is sort of hard to notice.

Modified:
ContestCounty.cs   <-- Methods, ResponseValue bag.
ContestCounty.hbm.xml  <-- ResponseValue bag.
ElectionContest.cs  <-- Methods
Response.cs  <-- Methods
ResponseValue  <-- Methods
frmAbstractReport.cs  <-- Reports
frmContestVoteSumry.cs  <-- Reports