Fixed issue with printing reports not honoring the default page margins set by the application.
The symptom is that the first print preview appears to be ok, but then the initial print is not the same, in my case shifted 1/2 inch to the right.  It appears the problem is not using the margins established when the document was created.  The act of printing then updated the margin settings on the page setup dialog and the print preview.